Move Peephole to separate file
